third_party/WebKit/LayoutTests/external/wpt/dom/nodes/Document-constructor-expected.txt

This is a testharness.js-based test.
PASS new Document(): interfaces
PASS new Document(): children
PASS new Document(): metadata
PASS new Document(): characterSet aliases
FAIL new Document(): URL parsingassert_equals: expected "http://example.org/?%C3%A4" but got "http://example.org/?ä"
Harness: the test ran to completion.
